requirements.txt
-r requirements-dev.txt
gunicorn
psycopg2
boto==2.48.0
boto3==1.7.4
botocore==1.10.4
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-allauth==0.34.0
django-author==1.0.1
django-rest-framework==0.1.0
django-s3-upload==0.1.2
django-storages==1.6.6
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
setuptools-git==1.2
six==1.11.0
static3==0.7.0
urllib3==1.22
